fetch/metadata/generated/element-area.https.sub.tentative.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/element-area.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title>HTTP headers on request for HTML "area" element navigation</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(url, {test, userActivated, attributes}) {
    const win = window.open();
    const area = win.document.createElement('area');
    area.setAttribute('href', url);

    for (const [ name, value ] of Object.entries(attributes)) {
      area.setAttribute(name, value);
    }

    win.document.body.appendChild(area);

    test.add_cleanup(() => win.close());

    if (userActivated) {
      test_driver.bless('enable user activation', () => area.click());
    } else {
      area.click();
    }
  }

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      makeRequestURL(key, ['httpsOrigin'], {mime: 'text/html'}),
      {
        test: t,
        userActivated: false,
        attributes: {}
      }
    );

    // `induceRequest` does not necessarily trigger a navigation, so the Python
    // handler must be polled until it has received the initial request.
    return retrieve(key, {poll: true})
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-frame-ancestors');
          assert_array_equals(headers['sec-fetch-frame-ancestors'], ['same-origin']);
        });
  }, 'sec-fetch-frame-ancestors - Same origin - no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      makeRequestURL(key, ['httpsCrossSite'], {mime: 'text/html'}),
      {
        test: t,
        userActivated: false,
        attributes: {}
      }
    );

    // `induceRequest` does not necessarily trigger a navigation, so the Python
    // handler must be polled until it has received the initial request.
    return retrieve(key, {poll: true})
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-frame-ancestors');
          assert_array_equals(headers['sec-fetch-frame-ancestors'], ['cross-site']);
        });
  }, 'sec-fetch-frame-ancestors - Cross-site - no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(
      makeRequestURL(key, ['httpsSameSite'], {mime: 'text/html'}),
      {
        test: t,
        userActivated: false,
        attributes: {}
      }
    );

    // `induceRequest` does not necessarily trigger a navigation, so the Python
    // handler must be polled until it has received the initial request.
    return retrieve(key, {poll: true})
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-frame-ancestors');
          assert_array_equals(headers['sec-fetch-frame-ancestors'], ['same-site']);
        });
  }, 'sec-fetch-frame-ancestors - Same site - no attributes');
  </script>
  </body>
</html>
